Move GraphicsSettings to Framework; fix compilation errors
