--- DX12Renderer/src/GraphicsSettings.cpp
@@ -1,1 +1,0 @@
-﻿#include "GraphicsSettings.h"
