+++ Framework/include/Framework/GraphicsSettings.h
